Rename the component
